AI-Resume-Screener-main/requirements.txt
Django
PyPDF2
python-docx
scikit-learn
nltk
pandas
sentence-transformers
numpy
spacy
